breakout mod Shrink adaptor iterator
